Implemented dropdown for project leads

+++ web_ama/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect,get_object_or_404
 from .forms import *
 
 from django.contrib.auth.models import auth
@@ -47,7 +47,20 @@ from django.utils import timezone
 from dateutil.relativedelta import relativedelta
 from django.db.models import Q, Count, Avg
 from django.core.paginator import Paginator
+from django.shortcuts import redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
+@login_required(login_url='my_login')
+def delete_record(request, pk):
+    record = get_object_or_404(process_assessment, id=pk)
+    
+    if request.method == 'GET':
+        record.delete()
+        messages.success(request, "Record deleted successfully.")
+        return redirect('dashboard')
+    
+    return redirect('dashboard')
 
 def home(request):
     return render(request,'web_ama/index.html')
@@ -129,7 +142,7 @@ def get_filtered_assessments(request):
     items_per_page = int(request.GET.get('items_per_page', 10))
     page_number = int(request.GET.get('page', 1))
 
-    assessments = process_assessment.objects.all()
+    assessments = process_assessment.objects.all().order_by('-creation_date')
 
     if search_term:
         assessments = assessments.filter(
@@ -162,7 +175,7 @@ def get_filtered_assessments(request):
         'process_division': str(assessment.process_division),
         'process_area': str(assessment.process_area),
         'process_description': assessment.process_description,
-        'process_lead': assessment.process_lead,
+        'process_lead': str(assessment.process_lead),
         'process_status': assessment.process_status,
         'process_name_c': assessment.process_name_c,
         'process_name_s': assessment.process_name_s,
@@ -342,14 +355,14 @@ def update_record(request, pk):
 
 # - View a singular record
 
-def format_li(line):
-    """Format the proposals as bullet points."""
-    lines = line.split('\n')
-    return [line.strip() for line in lines if line.strip()]
 
 @login_required(login_url='my_login')
 def singular_record(request,pk):
-
+    
+    def format_li(line):
+        """Format the proposals as bullet points."""
+        lines = line.split('\n')
+        return [line.strip() for line in lines if line.strip()]
     process = process_assessment.objects.get(id=pk)
 
 
@@ -369,12 +382,15 @@ def singular_record(request,pk):
 
 # - Delete a record
 @login_required(login_url='my_login')
-def delete_record(request,pk):
-    record = process_assessment.objects.get(id=pk)
-
-    record.delete()
+def delete_record(request, pk):
+    record = get_object_or_404(process_assessment, id=pk)
+    
+    if request.method == 'GET':
+        record.delete()
+        messages.success(request, "Record deleted successfully.")
+        return redirect('dashboard')
+    
     return redirect('dashboard')
-
 
 
 # - Generate a PDF
